relocate and rename run-id env var
