+++ cstar/orchestration/models.py
@@ -1,4 +1,5 @@
 import itertools
+import os
 import typing as t
 from abc import ABC
 from copy import deepcopy
@@ -23,9 +24,9 @@ from pydantic import (
 )
 from pytimeparse import parse
 
+from cstar.base.env import ENV_CSTAR_RUNID
 from cstar.base.utils import slugify
 from cstar.execution.file_system import RomsFileSystemManager
-from cstar.orchestration.utils import get_run_id
 
 RequiredString: t.TypeAlias = t.Annotated[
     str,
@@ -502,7 +503,7 @@ class Step(BaseModel):
         Path
             The path to the step working directory.
         """
-        run_dir = get_run_id()
+        run_dir = os.environ[ENV_CSTAR_RUNID]
         step_dir = self.safe_name
         return Path(run_dir) / step_dir
