+++ cstar/orchestration/utils.py
@@ -1,16 +1,9 @@
-import os
 import typing as t
-from datetime import datetime, timezone
 
 from cstar.base.env import EnvVar
 
 _GROUP_ORCH: t.Final[str] = "Orchestration"
 _GROUP_DEV: t.Final[str] = "Developer Only"
-
-
-def generate_run_id() -> str:
-    """Generate a unique run identifier based on the current time."""
-    return datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
 
 
 ENV_CSTAR_ORCH_DELAYS: t.Annotated[
@@ -43,16 +36,6 @@ ENV_CSTAR_CMD_CONVERTER_OVERRIDE: t.Annotated[
 ] = "CSTAR_CMD_CONVERTER_OVERRIDE"
 """Environment variable containing an overridden mapping key to apply when
 converting applications into CLI commands."""
-
-ENV_CSTAR_ORCH_RUNID: t.Annotated[
-    t.Literal["CSTAR_RUNID"],
-    EnvVar(
-        description="Unique run identifier used by the orchestrator.",
-        group=_GROUP_DEV,
-        default_factory=lambda _: generate_run_id(),
-    ),
-] = "CSTAR_RUNID"
-"""Environment variable containing a unique run identifier used by the orchestrator."""
 
 ENV_CSTAR_SLURM_ACCOUNT: t.Annotated[
     t.Literal["CSTAR_SLURM_ACCOUNT"],
@@ -93,15 +76,3 @@ ENV_CSTAR_ORCH_REQD_ENV: t.Annotated[
     ),
 ] = "CSTAR_ORCH_REQD_ENV"
 """A comma-delimited list of required env configuration values. TEMPORARY (move to CStarEnvironment / per-platform settings?)."""
-
-
-def get_run_id() -> str:
-    """Retrieve the current run-id.
-
-    Generate a new run-id if not found in the environment.
-
-    Returns
-    -------
-    str
-    """
-    return os.getenv(ENV_CSTAR_ORCH_RUNID, generate_run_id())
