+++ cstar/tests/unit_tests/orchestration/test_orchestration.py
@@ -8,7 +8,7 @@ from unittest import mock
 import networkx as nx
 import pytest
 
-from cstar.base.env import FLAG_ON
+from cstar.base.env import ENV_CSTAR_RUNID, FLAG_ON
 from cstar.base.feature import ENV_FF_ORCH_TRX_TIMESPLIT
 from cstar.orchestration.launch.local import LocalLauncher
 from cstar.orchestration.models import Application, RomsMarblBlueprint, Step, Workplan
@@ -26,7 +26,6 @@ from cstar.orchestration.transforms import (
     WorkplanTransformer,
     get_time_slices,
 )
-from cstar.orchestration.utils import ENV_CSTAR_ORCH_RUNID
 
 
 @pytest.fixture
@@ -319,7 +318,7 @@ def test_workplan_transformation(diamond_workplan: Workplan) -> None:
         mock.patch.dict(
             os.environ,
             {
-                ENV_CSTAR_ORCH_RUNID: str(uuid.uuid4()),
+                ENV_CSTAR_RUNID: str(uuid.uuid4()),
                 ENV_FF_ORCH_TRX_TIMESPLIT: FLAG_ON,
             },
         ),
